refactor(frontend): split the Home NeedsYou tile into sub-components
Extract the review decision workflow into useAttentionDecisions, the tile
header into NeedsYouTitle and the approve/decline controls into
AttentionRowActions, bringing NeedsYou.tsx, AttentionRow.tsx and
useNeedsYou.ts back under the ~50 line render/hook guideline.

Also give HomeTileSkeleton a named Props interface instead of an inline
props type.
